refactor: move parser into new folder
