+++ types/parser/index.d.ts
@@ -1,0 +1,8 @@
+export declare function parseArguments(rawArgs: any): {
+    createConfig: boolean;
+    skipPrompt: boolean;
+    help: boolean;
+    project: {
+        version: boolean;
+    };
+};
